:sparkles: Create reading-specs theme
This is a complimentary theme intended to be used over standardized color
themes, like Zenburn, Molokai and so on.
This way I can use the fonts I like over these themes without pushing changes or
having to customize their code. In this last case, my changes could be
overridden when the themes get new versions.