Adjust some org-mode snippets
Now that I know some Elisp-fu, I can correct these snippets. They
weren't working properly before, because the text didn't show up in the
middle of the delimiters.

With yas-moving-away-p, we can ensure the text will be fully formed
before enclosing it in the delimiters. I'd like to have a faster
feedback cycle on this, but it works for now.