Include Smartparens
This is the big gun delimiter solution for Emacs.
While Auto-pair mode just adds the delimiters you want, Smartparens does
that and much more.

We need it for Elixir because of the support for the language. Auto-pair
does not add a do-end pair automatically and we can't just use
ruby-electric mode. It works for most cases, but Elixir's syntax differs
in a couple points.