Create apib-mode tweaks
Originally, apib mode inherited the customizations made in
markdown-mode, since it is derived from that mode, but the use case for
them is quite different.

In an APIb file, you often need to know in which line you're in and
Emacs should not wrap lines. The monospace font is necessary for
aligning JSON snippets and format the APIb file properly.