Adjust projectile-phoenix config
This was also a problem in projectile-rails projects: you could use the
package's bindings only in Ruby/Elixir buffers.

With this change, the bindings will be available for every file inside
the project, which is what the user wants most of the time, and the goal
of these packages is to facilitate navigation between the framework's
files and commands.