Bring back ace-jump mode
I brought this back because now I have an use case for it. I'm using a
lot of windows and different frames and this makes it possible to switch
directly to the section of the other buffer where I want to do stuff,
instead of switching windows and then going to where I want.