Add display-buffer-alist customizations
This variable can hold rules for where Emacs should display the buffers
in the current frame (or even other frames). With this the buffers
always show up in the same place, and you can even preserve the window
layout you prefer without resorting to stuff like desktop-save-mode or a
persistent session.