Reuse same occur/search window in Emacs state
These search buffers were made to be used with Emacs keybindings, so
there's not much point in using Evil keys there.

We also adjusted `display-buffer-alist` to reuse the same search buffer
window that was already created in the first search when hitting g
there, which is the keybinding to redo that search and display the new
results.

You can use these search buffers as a running list of code points of
interest when reviewing or changing code.