Use Olivetti mode and Helvetica font for writing
Helvetica just looks so much better than the other fonts for me. Yes,
this is one of the things that I miss from using Macs, they know how to
design their visual stuff.

Olivetti-mode is one of those packages that you don't use often, but you
miss it when you need it. Especially for writing long-form text and documentation.