Don't auto-save files in text-mode derivatives
It's annoying to have Emacs auto-save Org-like files and clean out all
the whitespace and move my cursor when I'm in the middle of formulating
my thoughts and thinking without typing at the buffer.

We could also put a condition on the function to clear the whitespace to
skip running when auto-saving, but I think this is much easier, since I
have Git and version my Org files. There's also the local Emacs backups
in case something bad happens.