Adjust fonts for pure GTK configuration
The current version of Emacs can be compiled using pure GTK.

This brings all the improvements from GTK itself to it, but it doesn't
identify multiple frames, so functions like `other-frame` will not
behave properly.

This is not a problem if you use Emacs with a single display, but it is
a problem if you use multiple displays. That is the case at work for me,
so I need to split my configuration like this.