Improve performance on auralcat/magit-open-changed-files-from-main
Opening the files in Emacs with `find` instead of passing a list of
files to `emacsclient` doesn't block Emacs, but it makes the buffers be
opened one by one immediately.

This may be what you want, but if you find you don't want that, you
should tweak the function to open the buffers with `find-file-noselect`.