Create functions to copy code to GFM blocks
I feel this is one of the basic things that we take for granted in
Emacs, that it can do _anything_ with plain text, but we never get to
use or tweak it to our needs.

In my case, I'm writing an awful lot of code in comments to explain my
reasoning for comments in pull requests, and copying code from Emacs (or
worse, GitHub itself), writing the `s to make the fenced code blocks
there and pasting the code you want to share.

There is also the option to share the link to the code in the current
branch (in GitHub, of course), but I think that adds an extra layer of
friction that we don't need when explaining things: the user has to
click the link to see the code instead of reading it right there. The
upside of this approach is that you don't add a wall of code in the
middle of the text, so there are tradeoffs in both approaches.

Side note: writing fenced code blocks by hand is a pain when you are
using a non-English keyboard layout, and even then I think this could be
improved with at least a general code block shortcut. Currently GitHub
only offers the `suggestion` code block shortcut with M-g when writing
comments in pull requests, then you'd need to change the language name
in the code block to add other kinds of syntax there.

This is why I came up with these functions to add the text I'm copying
from modes derived from prog-mode (which correspond to programming
language modes) into a fenced code block depending on what I need to do:

- Just some example: copy it to a named GFM code block
- Highlight something important: copy it to a diff GFM code block
- Want to add a suggestion to the pull request from the comfort of my
  Emacs: copy it to a suggestion GFM code block
- GFM syntax highlighting is not supported where I'm pasting this code:
  use the plain Markdown code block.

Depending on how I use this, I think these functions could evolve into a
package! But let's try them out a bit first and see how to go from there.