:fire: Remove vterm
I'm using tmux in a dedicated terminal nowadays instead.

It started with terminal modes in Emacs being slower to output text from
`mix test`. Later on I needed to work with multiple services and they
needed their own session for that, so I switched to tmux in a dedicated
terminal emulator.