Add termint package
This is a new package that allows you to run comint-mode buffers in a vterm
terminal or eat!
This improves the experience in Emacs by 1000%!
And yes, we can even SSH into other servers and stuff using a standard terminal
emulator with all the color gloriousness and from the comfort of your Emacs!